refactor: move `main.gui` into its own module
